Separate set_status_bonuses() from set_creature_bonuses().
git-svn-id: http://svn.osdn.net/svnroot/deeangband@3482 89dd94fa-70ca-40ae-8747-2094edff379b

+++ source/trunk/src/types.h
@@ -1628,7 +1628,7 @@ struct creature_type
 
 	u32b change_floor_mode;
 	s16b time_stopper;
-	int tsuri_dir;
+	byte tsuri_dir;
 };
 
 // creature_flags_type
